Updated user entity

+++ src/main/java/com/bitwiseglobal/resumemgmt/repository/IUserRepository.java
@@ -1,9 +1,11 @@
 package com.bitwiseglobal.resumemgmt.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 
 import com.bitwiseglobal.resumemgmt.entityvo.User;
 
-public interface IUserRepository extends CrudRepository<User, Long>{
+public interface IUserRepository extends CrudRepository<User, BigInteger>{
 
 }
